+++ mythic_container/utils_mythic_file_transfer.py
@@ -18,6 +18,9 @@ async def sendFileToMythic(contents: bytes, agentFileId: str) -> bool:
                     else:
                         logger.error(f"[-] Failed to upload file contents to Mythic")
                         return False
+                else:
+                    logger.error(f"[-] Failed to upload file contents to Mythic")
+                    return False
     except Exception as e:
         logger.exception(f"[-] Failed to upload payload contents: {e}")
         return False
@@ -31,6 +34,9 @@ async def getFileFromMythic(agentFileId: str) -> bytes:
                 if resp.status == 200:
                     responseData = await resp.read()
                     return responseData
+                else:
+                    logger.error(f"[-] Failed to download file from Mythic: {resp.status}")
+                    raise Exception("[-] Failed to download file from Mythic")
     except Exception as e:
         logger.exception(f"[-] Failed to upload payload contents: {e}")
-        return None
+        raise Exception("[-] Failed to download file from Mythic")